refactor(car name validator): CarName 클래스에서 이름 검증하지 않고 CarNameValidator에서 검증하도록 수정
