+++ src/main/java/racingcar/model/RandomNumberGenerator.java
