--- src/main/java/racingcar/validator/Validator.java
@@ -1,6 +1,0 @@
-package racingcar.validator;
-
-public interface Validator {
-    int convertTryCountToInt(String stringTryCount);
-    void validTryCount(int tryCount);
-}
